inspection_robot/requirements.txt
PyQt5==5.15.9
pyqtgraph==0.13.3
numpy==1.24.3
scipy==1.11.1
matplotlib==3.7.2
pandas==2.0.3
pyserial==3.5
Pillow==10.0.0

